+++ tests/unit/mixer-dataset.js
@@ -14,7 +14,7 @@ chai.use(require('chai-as-promised'));
 
 describe('mixitup()', () => {
     it('should throw an error if `load.dataset` does not match pre-rendered targets', () => {
-        let emptyContainer = dom.getEmptyContainer();
+        const emptyContainer = dom.getEmptyContainer();
 
         chai.assert.throws(() => {
             mixitup(emptyContainer, {
@@ -26,7 +26,7 @@ describe('mixitup()', () => {
     });
 
     it('should throw an error if UID not provided in dataset API mode', () => {
-        let container = dom.getContainer();
+        const container = dom.getContainer();
         let mixer;
 
         chai.assert.throws(() => {
@@ -39,10 +39,10 @@ describe('mixitup()', () => {
     });
 
     it('should instantiate in dataset API mode when provided with `load.dataset` and a matching container', () => {
-        let container = dom.getContainer();
-        let targets = Array.prototype.slice.call(container.children);
+        const container = dom.getContainer();
+        const targets = Array.prototype.slice.call(container.children);
 
-        let mixer = mixitup(container, {
+        const mixer = mixitup(container, {
             data: {
                 uidKey: 'id'
             },
@@ -51,7 +51,7 @@ describe('mixitup()', () => {
             }
         });
 
-        let state = mixer.getState();
+        const state = mixer.getState();
 
         chai.assert.equal(state.activeFilter, null);
         chai.assert.equal(state.activeSort, null);
@@ -66,9 +66,9 @@ describe('mixitup()', () => {
 
 describe('mixitup.Mixer', () => {
     describe('#dataset()', () => {
-        let container = dom.getContainer();
-        let workingDataset = dataset.slice();
-        let config = {
+        const container = dom.getContainer();
+        const workingDataset = dataset.slice();
+        const config = {
             data: {
                 uidKey: 'id',
                 dirtyCheck: true
@@ -81,16 +81,16 @@ describe('mixitup.Mixer', () => {
             }
         };
 
-        let mixer = mixitup(container, config);
+        const mixer = mixitup(container, config);
 
-        let startTotalWhitespace = dom.getTotalWhitespace(container.outerHTML);
+        const startTotalWhitespace = dom.getTotalWhitespace(container.outerHTML);
 
         after(() => mixer.destroy());
 
         it('should throw an error if an item is added to the dataset, without a render function defined', () => {
-            let newDataset = dataset.slice();
-            let container = dom.getContainer();
-            let erMixer = mixitup(container, {
+            const newDataset = dataset.slice();
+            const container = dom.getContainer();
+            const erMixer = mixitup(container, {
                 data: {
                     uidKey: 'id'
                 },
@@ -110,10 +110,10 @@ describe('mixitup.Mixer', () => {
         });
 
         it('should throw an error if an item is added to the dataset without a valid UID', () => {
-            let newDataset = dataset.slice();
-            let container = dom.getContainer();
+            const newDataset = dataset.slice();
+            const container = dom.getContainer();
 
-            let erMixer = mixitup(container, config);
+            const erMixer = mixitup(container, config);
 
             newDataset.push(new dom.Item({
                 categories: ['d']
@@ -127,20 +127,20 @@ describe('mixitup.Mixer', () => {
         });
 
         it('should throw an error if an item with a duplicate UID is added to the dataset', () => {
-            let newDataset = dataset.slice();
-            let container = dom.getContainer();
+            const newDataset = dataset.slice();
+            const container = dom.getContainer();
 
-            let erMixer = mixitup(container, config);
+            const erMixer = mixitup(container, config);
 
             newDataset.push(new dom.Item({
-                id: 1,
+                id: 'target-1',
                 categories: ['d']
             }));
 
             chai.assert.throws(() => {
                 erMixer.dataset(newDataset);
             }, mixitup.messages.errorDatasetDuplicateUid({
-                uid: '1'
+                uid: 'target-1'
             }));
         });
 
@@ -207,7 +207,7 @@ describe('mixitup.Mixer', () => {
         });
 
         it('should remove a target when an item is removed from an arbitary point in the dataset', () => {
-            let removed = workingDataset.splice(2, 1);
+            const removed = workingDataset.splice(2, 1);
 
             chai.assert.equal(removed[0].id, 999);
 
@@ -221,11 +221,11 @@ describe('mixitup.Mixer', () => {
         it('should sort targets when the dataset is sorted', () => {
             workingDataset.reverse();
 
-            let ids = workingDataset.map((item) => item.id.toString());
+            const ids = workingDataset.map((item) => item.id.toString());
 
             return mixer.dataset(workingDataset)
                 .then((state) => {
-                    let elIds = state.show.map((el) => el.id);
+                    const elIds = state.show.map((el) => el.id);
 
                     chai.assert.equal(state.totalShow, 6);
                     chai.assert.deepEqual(ids, elIds);
@@ -250,15 +250,15 @@ describe('mixitup.Mixer', () => {
         it('should accept a callback function which is invoked after dataset change', () => {
             workingDataset.reverse();
 
-            let ids = workingDataset.map((item) => item.id.toString());
+            const ids = workingDataset.map((item) => item.id.toString());
 
-            let promise = new Promise(resolve => mixer.dataset(workingDataset, resolve));
+            const promise = new Promise(resolve => mixer.dataset(workingDataset, resolve));
 
             chai.assert.isFulfilled(promise);
 
             return promise
                 .then((state) => {
-                    let elIds = state.show.map((el) => el.id);
+                    const elIds = state.show.map((el) => el.id);
 
                     chai.assert.equal(state.totalShow, 6);
                     chai.assert.deepEqual(ids, elIds);
@@ -268,11 +268,11 @@ describe('mixitup.Mixer', () => {
         it('should accept a boolean allowing toggling off of animation', () => {
             workingDataset.reverse();
 
-            let ids = workingDataset.map((item) => item.id.toString());
+            const ids = workingDataset.map((item) => item.id.toString());
 
             return mixer.dataset(workingDataset, false)
                 .then(state => {
-                    let elIds = state.show.map((el) => el.id);
+                    const elIds = state.show.map((el) => el.id);
 
                     chai.assert.equal(state.totalShow, 6);
                     chai.assert.deepEqual(ids, elIds);